+++ src/test/converter2/issues/gh3020.ts
@@ -1,0 +1,7 @@
+/**
+ * Component demo.
+ *
+ * @fires {CustomEvent<{id: string, source: Element}>} item-click - event when item is clicked.
+ */
+export class ButtonControlElement extends Object {
+}